Implemented TimingPoints and (almost) HitObjects reading

+++ Enums/Effects.cs
@@ -1,7 +1,7 @@
 ﻿namespace OsuFormatReader.Enums;
 
 [Flags]
-public enum Effects : byte
+public enum Effects : int
 {
     KiaiTime = 1 << 0,
     FirstBarLineOmittion = 1 << 3
